requirements.txt
# Core Flask web framework
Flask>=2.3.0
Jinja2>=3.1.2

# Production WSGI server
gunicorn>=21.0.0

# Data collection and processing
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Reddit API
praw>=7.7.0

# Google APIs and Trends
google-api-python-client>=2.90.0
pytrends>=4.9.0

# Twitter API
tweepy>=4.14.0

# Text analysis
textblob>=0.17.1

# Data processing core
numpy>=1.24.0

# Scheduling
APScheduler>=3.10.0

# Date utilities
python-dateutil>=2.8.2

# Logging utilities (Windows compatibility)
colorama>=0.4.6

# Web scraping for Upwork (fixes distutils error)
selenium>=4.15.0
undetected-chromedriver>=3.5.0
setuptools>=68.0.0 